+++ include/unistd.h
@@ -72,5 +72,6 @@ void _exit(int status);
 int open(const char *path, int flags, ...);
 int close(int fd);
 int dup2(int oldfd, int newfd);
+int isatty(int fd);
 
 #endif