+++ lib/libc/sys/termios.c
@@ -1,8 +1,6 @@
 /*-
- * Copyright (c) 2002-2018 The UbixOS Project.
+ * Copyright (c) 2002-2026 The UbixOS Project.
  * All rights reserved.
- *
- * This was developed by Christopher W. Olsen for the UbixOS Project.
  *
  * Redistribution and use in source and binary forms, with or without modification, are permitted
  * provided that the following conditions are met:
@@ -26,38 +24,61 @@
  * OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef _UBIXOS_TTY_H
-#define _UBIXOS_TTY_H
+#include <termios.h>
+#include <sys/ioccom.h>
 
-#include <sys/types.h>
+int
+tcgetattr(int fd, struct termios *t)
+{
+	return (ioctl(fd, TIOCGETA, t));
+}
 
-#define TTY_MAX_TERMS 5
+int
+tcsetattr(int fd, int action, const struct termios *t)
+{
+	unsigned long req;
 
-/* tty_setmode cmd values */
-#define TTY_SETRAW   0  /* val 1 = raw, 0 = canonical */
-#define TTY_SETECHO  1  /* val 1 = echo on, 0 = echo off */
+	switch (action) {
+	case TCSADRAIN:	req = TIOCSETAW; break;
+	case TCSAFLUSH:	req = TIOCSETAF; break;
+	default:	req = TIOCSETA;  break;
+	}
+	return (ioctl(fd, req, (void *)t));
+}
 
-typedef struct tty_termNode {
-    char *tty_buffer;
-    char *tty_pointer;
-    uint8_t tty_colour;
-    uint16_t tty_x;
-    uint16_t tty_y;
-    pidType owner;
-    char stdin[512];
-    int stdinSize;
-    /* Line discipline */
-    char t_linebuf[512]; /* canonical input buffer (ISR fills until Enter) */
-    int  t_linelen;      /* chars currently in t_linebuf */
-    uint8_t t_echo;      /* 1 = echo input to terminal (default) */
-    uint8_t t_raw;       /* 1 = raw mode: bypass line discipline */
-} tty_term;
+void
+cfmakeraw(struct termios *t)
+{
+	t->c_iflag &= ~(IMAXBEL | IXOFF | IXON | ICRNL);
+	t->c_oflag &= ~OPOST;
+	t->c_lflag &= ~(ECHO | ECHOE | ECHOK | ECHOKE | ECHONL | ECHOPRT |
+	    ECHOCTL | ICANON | ISIG | IEXTEN);
+	t->c_cc[VMIN]  = 1;
+	t->c_cc[VTIME] = 0;
+}
 
-int tty_init();
-int tty_change(uInt16);
-tty_term *tty_find(uInt16);
-int tty_print(char *, tty_term *);
+int
+cfsetispeed(struct termios *t, speed_t speed)
+{
+	t->c_ispeed = speed;
+	return (0);
+}
 
-extern tty_term *tty_foreground;
+int
+cfsetospeed(struct termios *t, speed_t speed)
+{
+	t->c_ospeed = speed;
+	return (0);
+}
 
-#endif
+speed_t
+cfgetispeed(const struct termios *t)
+{
+	return (t->c_ispeed);
+}
+
+speed_t
+cfgetospeed(const struct termios *t)
+{
+	return (t->c_ospeed);
+}